inventory_browser: extract controls to their own component
